refactor: rename vk.rs -> proof.rs
